frontend: Revoke token modal errors
fixes: #754

Fix issue when `kubectl-client-id` is not provided. In this case, app
will not provide option to revoke refresh token. If no options are
available for the clien cog, then the cog is not displayed. If the user
has no `displayName` registered, then a placeholder is shown that is
consistent with the user page default.