frontend: Create React version of <tags-input> and use on Search page
Also migrates more of the search page functionality from Angular to
React so that only the URL rewriting logic is still in Angular. Note
that there are still some React warnings because that logic is still in
Angular. These warnings should be resolved when we migrate the URL
rewriting with the router migration to React.

Changes from Angular behavior:
  - Fix handling of "<" and ">" selectors (Angular version threw error
    when these were entered)
  - If a selector tag specified in the "q" GET param was invalid, the
    Angular input would still render the invalid tag as if it were
    valid. The React input discards the invalid tag instead.