frontend: Add Actions dropdown menu for all resource Overview pages
Add Actions menu for all resource types that have cog menus on their
list pages. In all cases, the Actions menu contains the same actions as
the corresponding cog menu, except for omitting any Edit actions
(because on the Overview page, the user can just click the Edit tab
link).

As part of this, rowOfKind is changed to take an "actions" prop instead
of using a hardcoded list of cog menu actions.