frontend: Migrate search pages to React Router
Had to fix the SearchPage component to re-render on search page route
changes. This wasn't a problem before because Angular was causing the
page to be recreated when the route changed. SearchPage now uses the URL
params as the source of truth for the current search options and is
therefore now a stateless component. Also added a key attribute to
ensure the search pane is re-rendered when search options change.