contrib/federation-experiment/federate: fix ingress bug
This commit fixes an odd ingress bug that appears when the ingress
controller pod is killed and restarted. When that occurs, ingress for
the Tectonic Cluster is broken. The only way to get ingress working
again is to delete the `federation-apiserver` ingress resource created
by the federate tool. Once this is deleted, ingress works again. You can
re-run the federat tool and see ingress continues to work until the
ingress controller pod is restarted. This patch fixes this behavior.