improve performance rendering nodes ...
The k8s API may update the resource version of any object even when no field changes.
We previously relied on the versions to determine if we should re-render a given object.
This caused *extremely* poor performance rendering Nodes and Cluster Settings.
We now do a deep comparison as a last resort (for all k8s objects).
This exposed a different bug where Rows which were collapsed/expanded were not immediately re-rendered.