frontend: Add Pod link on Container page
Sam wanted breadcrumbs to be added to the Container page, consistent with other pages. The way breadcrumbs work in NavTitle needs to updated in order for this to be possible. Sam suggested adding this link back to the Pod for the time being.

Addresses https://jira.coreos.com/browse/CONSOLE-544