Cache API discovery
On initial console page load, use cached API discovery resources when
possible. This noticeably improves initial console load times. Wait
until discovery completes before requesting resources not in the cache.

We still perform discovery on load to refresh the cache, but don't block
when we have the models we need.

The cache is invalidated when the console version changes. This avoids
errors if the internal model representation changes between versions.