Reduce tile view pages chunk size, fix minor display issues
Reduces the chunk size be eliminating the use of _.intersection in the tile views (was adding about 80k) and updating to latest patternfly-react-extensions which contains some dependency fixes.

Removed overrides for the tile description fader as this was fixed in patternfly-react-extensions.

Updated the tile view pages to always be full page rather than jumping when filtering.