Monitoring: Always show not firing alerts with status "Not Firing"
This behavior was trying to be helpful by indicating that a different
alert generated by the same rule was Firing / Pending. However, it was
actually very confusing because that Firing / Pending alert will have
different labels from those currently visible in the UI. Better to just
show the alert as Not Firing.