frontend: Make env delete button accessible
Made env delete button into a real button so it's in the tab flow and can be triggered by hitting return on the keyboard. Made the drag and drop button into a button as well so the styling would be consistent, but it's not in the tab flow since you can't drag and drop with the keyboard.

Fixes https://jira.coreos.com/projects/CONSOLE/issues/CONSOLE-1079?filter=myopenissues