Monitoring: Add metric graphs to Alert and Rule details pages
Adds a `QueryBrowser` component, which plots a PromQL query as a line
graph along with controls for changing the graph to display the query
for any time range.

Adds the `QueryBrowser` to both the alert details page and the rule
details page. The alert graph shows the single metric for that alert and
the rule graph shows all metrics data for that rule.

Upgrade Plotly to version 1.44.4.

Move the "View in Prometheus UI" link from the top of the page to the
graph area.

Make some changes to the graph Base class.
  - Make the `title` property optional
  - Add the ability to invoke `fetch()` without enabling polling
  - Set a lower limit of 5 seconds on the graph refresh interval to
    prevent too frequent polling when viewing narrow time ranges.
  - Add optional `numSamples` property that defines how many data points
    will be plotted for each metric