Fix bug where `pf-c-form-control` styles weren't loading in prod
Fixes https://jira.coreos.com/browse/CONSOLE-1588

IMPORTANT:  this bug occured because tree shaking only applies to production and not dev.  In other words, all PF4 CSS is included in dev, but in prod tree shaking removes any CSS for PF4 components that are not in use. Going forward, any time a PF4 react component styles are being used without the component itself, we need to explicitly include the relevant SCSS file as done in this PR.