Monitoring: Fix graph zoom & tooltips
Fix several related problems with graph zooming and tooltips.

The interaction between graph zooming and tooltips caused several bugs.
This fixes them by implementing our own zoom.

Removes the need to use Victory Chart's `VictorySelectionContainer` and
`createContainer()`.

Fixes
  - Bug where tooltip stopped displaying after graph data was updated
  - Improve laggy UI when selecting an area of the graph
  - Allow zooming on a region of the graph with no data points
  - Prevent zooming down to a time span so small that the Prometheus API
    returns an error
  - Prevent some unnecessary re-renders of the graph by replacing
    `Graph`'s `metadata` and `onZoom` props with a single
    `containerComponent` prop, which is memoized
  - Hide tooltips when zooming because they can get in the way when
    selecting the zoom region

Limits the zoom feature to only zooming the X-axis. This is preferable
to the previous behaviour of only being able to select a rectangular
region of the graph. Y-axis zooming may be brought back in a future PR.

Change cursor to `ew-resize` to give a visual indication that the graph
can be zoomed.