Query Browser: Fix table values
The table values were displaying the last value of each series, but that
is not actually correct since a series may not continue until the
current time. Instead we need to get the latest values from the
Prometheus API's `query` endpoint.

Also fixes a bug where the results table was not cleared after the query
input was cleared. A "No datapoints found" message is now shown instead.

Also fixes a bug where the table data was being truncated in Firefox.

Fixes: https://jira.coreos.com/browse/MON-744
Fixes: https://jira.coreos.com/browse/MON-734
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1739380