Query Browser: Fix UI rapidly flipping between two different queries
Fixes a bug where the Query Browser could get stuck rapidly flipping
between two different queries, causing the table results to change and
lines in the graph to appear and disappear.

This bug was caused by the poller callback updating the `query` field in
Redux, which under some circumstances caused the poller to be
reinitialized, triggering a loop.

The poller callback was updating `query` just for the Alert details page
and Alerting Rule details pages, so those pages now set `query`
themselves.