Bug 1794754: correctly show schema information for all core resources
Core k8s resources that didn't have a static model defined were missing
schema information in the Explore page and YAML editor sidebar. This is
because we were using API group `core` internally and incorrectly using
the internal group when looking up the resource in the OpenAPI doc.

This changes console to not use `core` in when creating models from
discovery.