VM Utilization: show datapoints from the most recent VMI only
Considering the scenario of start->stop->start VM.

Prior this fix, Prometheus data were merged for both the former
and later VMI objects causing confusing user experience (misleading graph).

With this fix, the retrieved data are not older than age of
the existing VMI object.

This is a workaround for the issue with collision of names
for old vs. recent VMI same as for lack of possibility
to use the UUID within the corresponding prometheus metrics.