Bug 1948566: Fix issue where react-modal was removing Operator modal from the DOM
Make sure that the Community Operator warning modal closes *before* the Operator details modal is opened.

Here's what I suspect was happening:

- User selects community operator from OperatorHub
- Community operator warning modal is triggered
- react-modal appends the appropriate modal component to the `#modal-container` DOM element
- User clicks "continue"
- 'showCommunityOperators' callback is called, which triggers the PF modal to be appended to the same `#modal-container` DOM element
- `close` callback is called which removes both the community operator modal and the Operator details modal elements from the DOM.
- The Operator details modal state is now corrupt because the underlying DOM element was removed by a parallel React instance.
- Calling close before showCommunityOperators resolves the problem.

I suspect that the udpates in React v17 and/or react-modal v3.12 made it possible for react-modal to remove DOM elements that it wasn't able to see/touch before, which is why this wasn't an issue up until now.