Bug 1968525: Warning - Operator Details page duplicate keys
We noticed an issue on the gitops operator details page in the browser
console where React claims 2 elements on the page have the same key.
I've tracked the issue down to the service accounts for the gitops
operator. There are 3 service accounts with the same
`serviceAccountName` value of `gitops-operator`. See screenshots.

https://bugzilla.redhat.com/show_bug.cgi?id=1968525