Bug 1986699: Differentiate between operators for install
Console was showing the wrong install status for operators with the same display name but different catalog sources. We don't have a ton of data to work with, but adding the package catalog source and namespace to the subscription finding function seems to work.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1986699