Bug 1990024: Fix typo
I fixed the spelling of eligible. I also fixed the namespaces of some strings in the public folder that were using the console-app namespace. It was causing all of the strings in console-app to show up in the public folder. That is expected behavior (it prevents namespace conflicts for i18next), but may be confusing for devs.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1990024