Bug 1985356: Check for nonexistent CSVs in installed block
The installed hint block for operators wasn't checking to see if the operator had a CSV or not in the K8sWatch hook. As a result, sometimes it would show an incorrect CSV link if there was no CSV on the operator. I added some additional logic to check so we can route them to the subscription if this happens.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1985356