Bug 2046531: Remove schema subdir from dynamic plugin sdk
When you build the console in version 4.9 the subdirectory [...]/packages/console-dynamic-plugin-sdk/schema gets created. In 4.10 it's created under [...]/packages/console-dynamic-plugin-sdk/generated/schema. So if you "./build.sh" under 4.9 and then switch to 4.10 and do "./clean.sh" and "./build.sh" the build fails, because only the subdirectory "generated/schema" was removed and the one on the same level as the directory "generated".

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2046531