Eliminate use of lookaside cache and move to Cachito
[Legacy Container First Pipeline][1] will be deprecated on Jun 3, 2022. Images failing to migrate to the new source container workflow by this date will not have their sources shipped to customers and lose the compliance with legal requirements.

To move to the new source container workflow, we need to:
- Eliminate use of distgit lookaside cache, which is not supported by the new workflow.
- Use [Cachito][2] to track and provide npm dependencies.
- Use [fetch-artifacts-url.yaml][3] to fetch binaries artifacts during the build.

Due to a limitation in Cachito, we also need to give a pseudo name and version to the frontend package. AFAIK the version doesn't need to be unique for every build.
I am not familiar with the code base. Let me know if the change is legitimate.

[1]: https://docs.engineering.redhat.com/display/SP/Decommission+of+Legacy+Container+First+Pipeline
[2]: https://github.com/containerbuildsystem/cachito
[3]: https://osbs.readthedocs.io/en/osbs_ocp3/users.html#fetch-artifacts-url-yaml