OCPBUGS-11668: Fix kebab actions on Installed Operators page
Update the Installed Operators page to use the useAccessReview hook to determine whether to list all subscriptions or just those in the current namespace. For users who don't have permission to list all subscriptions, the page will now only attempt to list subscriptions in the current namespace, and will show the correct actions for operators the current user has permission to edit or uninstall.